feat: refactor HangmanGame by moving keyboard and win/loss modal to HangmanKeyboard component
